backend/pyproject.toml
[project]
name = "fitness-workout-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.129.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "sqlalchemy[asyncio]>=2.0.46",
]
